+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 ## 0.10 *(2020-01-12 --- )*
 
+### Breaking
+ * `ValidateViolationsTask.action` is removed, override `processViolations` instead (#85).
+
 ### New
  * Gradle 5 compatible (up to 5.6.4) (#84)
  * Use Gradle 5.6.4 to build (#84)
@@ -13,6 +16,7 @@
  * Make sure nothing is logged when lint task is disabled.
 
 ### Internal
+ * Gradle validateTaskProperties issues (#85).
  * Test against latest minor versions of AGP (#96)
 
 ## 0.9 *(2019-02-19 --- 2019-07-11)*